Fix issue where dummy costs were recalculated each render
Previously if a re-render was caused in react, a new cost of the flight was created.
We want this to be consistent, so a property has been added to the flight data interface to hold cost. This is currently appended to the flight data as dummy data until it's pulled from the database.