Add function to register flight reservation with Reservations API endpoint
Adds function that posts to the reservation API endpoint  including the reservation data supplied as an argument to the API request body via JSON.

It will return the direct response or an error if the request could not be made.
If the request is made but has an error itself, the response will still be returned, allowing the function caller to do error handling with it.