Allow HomePage to use data of selected flight from FlightList
Previously when the flight was selected, that information was only known by the FlightList component which one was being used.
This means that the user could see which was selected, but the HomePage didn't have access to that data. So we couldn't use the flight's information for submission.

This updates the FlightList to accept a function property that is called when the flight selection changes, and it passes the newly selected flight as an IFlightData interface.

So when the selection changes, the HomePage has it and can use the price or other properties when registering the flight. The selected flight is now stored in the HomePage's state.