Submit button now registers flight submission with selected flight
Updates the submit button to use a new function that retrieves all the current filter data and flight data to submit the needed reservation to the reservations/ endpoint.

It then informs the user if the registration was a success, if there was an error, or if something is needed before a submission can occur.