Add image carousel component & banner images to home page
Adds new React component designed to display a set of images and cycle through them. This will be included a the top of the Home Page with both primary images and a background parallax image.

Updates webpack to handle importing jpg and png files.

Adds image assets that are imported.

Adds TypeScript typings so that jpg and png files can be imported as modules.

Updates the slogan styling again since the banner broke the text positioning.