Add dedicated search button & don't display submit button by default
We now have a search button displayed to the user to start the search instead of doing this dynamically.

Additionally, we don't want to have a submit button until we're ready for it to displayed. So getting the button is now delegated to a function that will be called to render the button.