HomePage now gets copy of multicity flight selections on update from MultiCityFlightSelection component
When the multi city flight selections in the MultiCityFlightSelections component change, a function passed into it from the HomePage.tsx as a prop is called. This allows us to pass up a copy of the changed flights array so that the homepage has all flights being used for the multicity mode.