Add ability for AirportSelector to be able to request locations data and store it as state
Adds the ability for the AirportSelector to be able to store locations data as an array in react state so that it can render the results as list items in an eventual dropdown menu of all selection choices for airports.

Adds a function that allows it to search the API endpoint with a given query to get relevant results from the input.