Adds awesome-debounce-promise to debounce API requests to prevent unnecessary traffic
Adds awesome-debounce-promise as an NPM dependency to add support for debouncing in React. This will allow us to prevent execution of an API call for a period of time so that we can check to see if we get another request in a period of time. If we do, the time will be extended, if not, we will fire the request.

Since we will be typing in a query and submitting an API request based on the typed value, we want to make sure we're not submitting an API call for every input, and should instead use debouncing to only submit the API request after the user has stopped typing their query. This will allow us to reduce the amount of API calls we're making significantly and still give a natural feel to searches being made.