Update AirportSelector to filter out non-airport Locations from endpoint
The locations endpoint includes locations that are non-airports and are simply cities such as "London". This is not useful for us, we don't want to select any cities or display any, so when we get the api request data full of location results, we will filter out non-airports before submitting the array to the react state.

This will allow us to give an empty array if there are no airports and still include all the actual airport data. This should allow us to not display these non-airports while still not introducing any odd behaviors in case there are no leftover results.