AirportSelector: Handle focus gain and loss & child element focus
Adds functions for on focus to inform the react state when our input component has gained focus and when it has lost focus. This is going to be used for determining when to and not to display the results list dropdown when performing a search.

However, input focus is not sufficient because onBlur will cause the dropdown to disappear even when clicking a list item from the dropdown, even though it's part of the component. So we need to check if the element being switched to is a child of the AirportSelector and not dismiss the dropdown in that scenario. (If we don't, the dropdown will close when clicking a list item, but the list item doesn't receive the click)