Update toast message to be closed manually instead of delay fade effect
The delay effect animation caused issues where it would not display again because the animation failed to reset on future displays when the component updated. Instead, the user will close the message by hand which will set the toast to hide. Then when it is set to show again, the message should show successfully.