Fix issue with Toast Type failing to re-open after first appearance
Fixed issue where state was not set if the props were different from state rather than if the previous props were different from the current state.

Added a function prop event that allows the parent component to know when the toast message has been closed so the show state can be updated from the top level. It needs to be set on the top level since it keeps feeding back into the child component.