Connected Number of Adult/Child travelers to react state & set default values in state
Updated the number of adult travelers input and the number of child travelers inputs to use react events that update the state of associated react state variables and pass in the values from state into the elements so they keep the values.

Also updated the default adult traveler value to be set in state so they're consistent. This will allow the element to update when the state default value is changed, and reduce risk of a conflict if one is updated and the other is not.