Add API handler for flightOffers/ endpoint and support for arguments
Adds a class that is intended to call the flightOffers/ API endpoint to pull in available flights based on filters passed into the API. This will allow us to get relevant flights for the filters we have selected.

Adds a interfaces to support all the associated types with the data passed in as well as an enum and interface for the arguments we should pass in or can pass in.