Update MultiCityFlightSelection component to use AirportSelector elements
Updates the MultiCityFlightSelection component to no longer use the default inputs for entering an airport and now uses the created AirportSelector component for retrieving list of airports and selecting one.

This gives significantly more information about the selection and more functionality to the user for finding the airport they want.