Add handling of multiple flights to search on Home Page (For all FlightType searches)
Adds state and helper functions for tracking which flights need to be searched and tracking search status. This will be used for selecting several flights in a row and seeing when we are done and ready to submit.

This will allow us to treat One Way, Round Trip, and Multi-City the same way and use the same logic for flight selection.