Add Loading Logo when waiting for results & ensure no data message shown if no results found
Adds a loading symbol to be displayed by the FlightList while waiting for data to load.
Add asynchronous handling to display the logo immediately and hide it once the data loads in. Either the data is displayed at that point, or if an error occurred it will pass an empty array so the flightlist knows to show a message that no data was foundl.