Ensure we get API data for current flight selection, avoid state refresh issues
Previously the first search would fail since the flight list needed to be assembled, but since we just set it in state, the state wasn't set by the time we tried to read the current flight.

Instead of waiting for another render, we'll pass the data directly to the API retrieval function. This is an optional argument, so if none is provided it will still read from state.