Update Search Progress when setting current searched flight as selected
Previously the search progress was not being updated when we set the current flight as selected so we could move to the next. This meant that the Flight List would still show we're selecting the first flight, and the Next button would still show instead of Submit if we were on the final page now.