No longer use useNavigate, just use Link and handle login by not showing actual link when not logged in
No longer uses useNavigate to change page since I never got it working with our class setup. Instead, the submit button is wrapped in a link that updates our reservations before changing page.

However, there are cases where we want to find out there is a problem before going to the checkout page, so we don't want to ALWAYS go to the checkout page.
Therefore, when the user is not logged in, they will see a dummy submit button that will just warn them to log in, where they will then get the real submit button that takes to the checkout page.