Display Login Button when not logged in during search to avoid confusing Submit button behavior
Previously the submit button would be present when the user was not logged in, however it would still take them the checkout page.

Then, a dummy submit button was created that warns the user they are not logged in when clicked until they login. This was also confusing and unclear, so a direct go to login button has been displayed in this case.