Update ImageCarousel to use CSS for image urls to avoid ReactRouter issues
There have been issues with Webpack and CSS-Loader converting relative urls while using ReactRouter. Multiple solutions failed to resolve the issue, so since relative URLs are working great, I have assigned each image to a style that can be used dynamically by the ImageCarousel component.

Not the ideal solution but it works and we don't have time to dwell on this aspect.